move out functions into their own modules
